+++ UI-autotests/src/test/java/helpers/Waiters.java
@@ -1,4 +1,4 @@
-package framework;
+package helpers;
 
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
