+++ UI-autotests/src/test/java/pages/UserHomePage.java
@@ -6,7 +6,7 @@ import org.openqa.selenium.support.ui.WebDriverWait;
 
 import java.time.Duration;
 
-import static framework.Waiters.waitUrlToBe;
+import static helpers.Waiters.waitUrlToBe;
 
 public class UserHomePage {
     private final WebDriver driver;
