+++ UI-autotests/src/test/java/tests/BaseTest.java
@@ -1,4 +1,4 @@
-package framework;
+package tests;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
