+++ UI-autotests/src/test/java/tests/LoginPageTests.java
@@ -1,5 +1,6 @@
+package tests;
+
 import data.TestsData;
-import framework.BaseTest;
 import io.qameta.allure.Epic;
 import io.qameta.allure.Feature;
 import io.qameta.allure.Owner;
