+++ UI-autotests/src/test/resources/META-INF/services/io.qameta.allure.listener.TestLifecycleListener
@@ -1,1 +1,1 @@
-framework.AllureListener
+helpers.AllureListener
